Use hashmap data structure
This commit changes the underlying data structure of the `geojson` and
`geojsonChanged` member variables. Instead of using an array a hashmap
is being used. As a key, the feature ID is used. This speeds up any
lookups and prevent the application from looping over potentially many
elements at the cost of consuming a little bit more memory.